Refactor code structure for improved readability and maintainability

+++ server/src/config/cloudinary.js
@@ -18,9 +18,11 @@ const uploadOnCloudinary = async (localFilePath) => {
     fs.unlinkSync(localFilePath);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
 
-    return error;
+    return { error };
   }
 };
